[FIX] account_sepa: warns users for non-Latin chars in SEPA Batch
When generating a SEPA XML file for a Batch Payment non-Latin characters
are currently removed from the XML. It is expected as SEPA only supports
latin characters. However, no warning is shown to the user which perturbs
them as they receive a modified XML (non-Latin chars are replaced by blanks).

Step to reproduce the issue:
1. Create an European company.
2. Install the Accounting module and install an European package
3. Go to Accounting > Configuration > Journals > Bank and set an account number
(e.g.: GB33 BUKB 2020 1555 5555 55)
4. Ensure that on the same page, the SEPA is available in the Outgoing Payments
5. Create a new vendor with Latin letters in its address.
6. Make sure that the new vendor has an Account Number in his Accounting tab
(e.g.: GB94 BARC 1020 1530 0934 59)
7. Go to Vendors > Payments and create one with Payment Method set to
"SEPA Credit Transfer" (ensure that Vendor Bank Account has been filled).
8. Go to Vendors > Batch Payments and create an Outbound SEPA payment. Select
the Payment we created in the previous step.
9. Generate the SEPA XML file.
You will see that in the address of the vendor, non-Latin characters have
been replaced by blanks without warning the user.

Solution: SEPA has released a mapping [1] to transform these non-Latin
characters into Latin ones but this mapping will not be directly included into
Odoo. This will be done by the R&D team and thus will take more time to develop
and is expected to arrive in later releases. In the meantime, a warning
message will just be shown to the user so that he knows that his SEPA XML
will be altered to fit the SEPA standards. A previous commit [2] made possible
to warns the user when generating the XML file.

[1] https://www.europeanpaymentscouncil.eu/document-library/guidance-documents/sepa-requirements-extended-character-set-unicode-subset-best
[2] https://github.com/odoo/enterprise/commit/b5d0420cb27aaef225c77b5f2b07657f8b9866a8

opw-2806383

closes odoo/enterprise#28060

X-original-commit: f7875e75e0a794b7c7f25d4907e84087672fc524
Signed-off-by: Olivier Colson <oco@odoo.com>
Signed-off-by: Desausoi Laurent (lade) <lade@odoo.com>