[FIX] helpdesk: allow users to list the ticket analysis
For the moment, there are no tree view for the Ticket Analysis model.
Consequently, when accessing the view, we only see the ID of the models, which
doesn't help the client.

Step to reproduce the issue:
1. Install the Helpdesk module
2. Go to Reporting > Ticket Analysis > Cohort View (last small symbol on the
top right)
3. Click on one of the entries to display the Ticket Analysis view
You should see only a list of IDs.

Solution: The Tree View related to the `helpdesk.ticket.report.analysis` did
not existed before. We had to create it.

opw-2839413

closes odoo/enterprise#29502

X-original-commit: 1303a55c856a0d6ce3e2b7edff3f513f3dc2d4f3
Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>
Signed-off-by: Desausoi Laurent (lade) <lade@odoo.com>