[FIX] l10n_cl_edi_boletas: daily sales book should not be copied on account.move
It was noticed while testing that by duplicating an accepted boleta it would
be added to an existing vat book in the draft state, while we never add it in
the draft state.  The link to the sales book was copied and that seemed to be
the culprit.  So, we put copy False.

closes odoo/enterprise#27460

X-original-commit: 825ed8eb051d3cbf4e2540755e1409b626aa6285
Signed-off-by: William André (wan) <wan@odoo.com>