[FIX] account_avatax{,_sale}: don't depend on _description
It's possible _description will change across versions and this would
cause mismatches between what's in Avatax and what's in Odoo. This
could cause reports to not be accurate and it could also cause
errors. E.g. in case an invoice is committed in one version and that
invoice is uncommitted in the new version (provided the _description
changed).

The provided _get_avatax_description implementations have copied the
previously used _description fields, so nothing should change.

Discussed in #26855 and related to task-2824272

closes odoo/enterprise#26947

X-original-commit: 3274798500c130ea2e3d21fe86fffea3635a8962
Signed-off-by: William André (wan) <wan@odoo.com>