[FIX] account: fix default journal filters applying to journal items
Since https://github.com/odoo/enterprise/pull/23903, when in the reports pages, having no journal checked would turn them all to checked.

This was coherent with the displayed report. However, selecting a single journal (for example) from having them all selected had become uncomfortable.

Furthermore and more importantly, the "All journals" filter would continue to be applied on audited journal items.

***Description of the bug***
Register two payments for two bills, one from the Bank account, the other from the Cash account.

On the Cash Flow Statement report, under 'Cash paid for operating activities', click 600000 Expenses --> Journal Items.

No line would appear because the journal filter would still be silently applied. It wasn't however possible to deactivate it from this page.

***Further informations discovered during the fixing***
This was the opportunity to realize that the informations displayed under the lines displayed in the audited journal items weren't coherent with the information we wish to display.

The filters applied are not pertinent and the lines in themselves are not always relevant to the situation.

Since this is error-prone and problematic, we decided to remove the access to those informations.

A bigger fix is planned and will make them available again, corrected.

***FIX***
Reverted the previous way of functioning concerning the Checking of Journals in the reports :
If no journal is selected, all of them still are taken into account in the reports, but they do not appear as checked.

This makes checking a few journals more comfortable and the filter doesn't transfer to another report's page afterwards.

Disabled the possibility to display the problematic journal items from the Cash Flow Report.

closes odoo/enterprise#27444

X-original-commit: afd2819d5afc271ef221ce30a69efe5d611e522d
Signed-off-by: Laurent Smet <las@odoo.com>
Signed-off-by: Camille Spiritus (casp) <casp@odoo.com>