[REF] documents: prepare do-action for new env
This PR helps reducing the noise in the one introducing the new environment in
the discuss app. Indeed, we won't rely on the bus anymore to listen to do actions,
we will instead use the action service. The diff is awfull because the indentation
has changed as well as the parameter names. Changing the indentation and the parameters
now will ease the review of the main task.

task-2582313

closes odoo/enterprise#26970

Related: odoo/odoo#90581
Signed-off-by: Sébastien Theys (seb) <seb@odoo.com>