[FIX] account_asset: auto_post move after asset modif°
With this commit, we set the asset's deprecation moves to
auto_post = True after modifying the asset.

Steps to reproduce:

- Create an asset for several months in the future, and confirm it
- Modify the asset
-> deprecation moves are not set to be auto post anymore

Use the _post method on moves newly created after the asset
modification.

A fix was already deployed with this commit 33998d754761f8fa31033414898e11bb86264190
but led to other issues, so we add a condition before calling _post().

opw-2823359
opw-2833078

closes odoo/enterprise#27121

X-original-commit: c4bac4e96176f757dd07fb4a379751fa2a5f0e13
Signed-off-by: William André (wan) <wan@odoo.com>