[FIX] documents_spreadsheet: support selection field in pivot formula
Group a pivot by any selection field and insert the pivot in the
spreadsheet.
=> Formulas with this field are in #ERROR

The field value (a string) is parsed as a number instead of a string

The code is refactored to make it more easily testable
I think we would go a bit further and "push" the
"raw" string values parsing up, such that the minimum amount
of code needs to handle that.

Task 2841206

closes odoo/enterprise#26980

Signed-off-by: Pierre Rousseau (pro) <pro@odoo.com>