[FIX] appointment: fix badge custom appointment type
When sharing a custom appointment type, if there are
slots on multiple months, we display a badge next to the
navigation arrows.
This badge contains the number of slots in the next/previous
months.

Before the badge didn't take into account the availability of
the slot. Now the number of slots match correctly with the
number of slots dipslayed.

task-2836723

closes odoo/enterprise#28340

X-original-commit: 9b6f8ef86078cecc23583ca213f9915551a299c6
Signed-off-by: Hennecart Jérémy (jeh) <jeh@odoo.com>
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>