[IMP] hr_contract_salary: adapt "generate offer" button
At the moment, if you are in Recrutment and you want to Generate an Offer,
if there is no contract template defined on the Job Position, you'll have an error message.

After this commit :
If the Contract Template is not defined on the Job Position, the wizard will open
but will display a warning message instead of the link.

TaskID #2837564

closes odoo/enterprise#27089

Signed-off-by: Kevin Baptiste <kba@odoo.com>