[FIX] project_enterprise: prevent variability in query counters.
Some tests that are run before those located in `project_enterprise/tests/test_auto_shift_dates.py`
seem to indeterministically cache information on groups `.` and `base.group_system`. This impacts
the variability of the query counters in the tests, which prevents to set them properly.

The purpose of this commit is to prevent this instability during the tests.

This commit reaches that goal by:

- Calling `has_group()` on both groups in the setUp of the tests so that it will be cached.
- Using `@users` to ensure that we are taking ir.rules into account (as the default __system__) won't.

closes odoo/enterprise#27678

X-original-commit: fc65592a69257efc4c8d84669d15e916cfe344f9
Signed-off-by: Xavier <xbo@odoo.com>
Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>