[FIX] account_bank_statement_import_camt: correct amount with charges
In case charges are specified at NtryDtls level, the corresponding Amt
element does not include the charges in the amount. In some files, this
element is sometimes not even specified, but only Amt element at Ntry
level is present. We should always look at this element in that case,
as it will always be present and include charges.

closes odoo/enterprise#27503

X-original-commit: 1824640a8eb27a38dfe4d31cb81c5b0f612033fe
Signed-off-by: Olivier Colson <oco@odoo.com>
Signed-off-by: Alex Thuyls (alt) <alt@odoo.com>