[FIX] l10n_co_edi: allow Credit/Debit notes creation without DIAN
We don't need the DIAN fields when we create a debit/credit note, however it is
mandatory when creating other invoices.

Step to reproduce the issue:
1. Install Accounting, l10n_co and l10n_co_edi modules
2. Go to Accounting > Customers > Credit Notes and create one
3. Save then Confirm
You will see that a pop-up is displayed to fill the "Resolucion DIAN" in the
appropriate Journal. These fields are in fact not required for Credit/Debit
Notes.

Solution: The condition concerning the DIAN fields has been updated to ignore
the case of Debit/Credit Notes.

opw-2780428

closes odoo/enterprise#27715

X-original-commit: 64d3068505659c545385891fb02de1f03ab6f14b
Signed-off-by: Josse Colpaert <jco@odoo.com>
Signed-off-by: Desausoi Laurent (lade) <lade@odoo.com>