[FIX] hr_payroll, l10n_*_hr_payroll: fix view on employer cost dashboard
Currently in the payroll app there is a boolean field responsible for showing
hr.salary.rules in the employer cost dashboard plot. However, some rules are
always shown, as they are hard-coded. This commit fixes this behavior by using
the boolean field to show these rules, therefore giving more flexibility to the
plots as we are able to hide the previous hard-coded rules.

task-2836922

closes odoo/enterprise#27743

X-original-commit: d90b7363afbd6a915dfac5a06270e3217a4630d6
Signed-off-by: Kevin Baptiste <kba@odoo.com>