[FIX] l10n_co_edi: clean the transaction id when error so you can button_draft
The problem is that you can not reset to draft when the
transaction_id is still there, which is logical as we
are still processing the transaction with Carvajal.

However, when we get an error from Carvajal, we do
not clean the transaction_id, while it should happen then
and not (only) when we reset to draft as at that time,
we can not do anything with it anymore anyways as it is
rejected.  And this way, we can also reset to draft at that
point.

opw-2838326

closes odoo/enterprise#27199

X-original-commit: fcbaa1d2f75e399b32cb908e763f153d66655968
Signed-off-by: William André (wan) <wan@odoo.com>