[IMP] Documents: Allow user-specific customization on list view
There isn't enough place on a default phone screen to display the 6
default fields of the documents list view*. Unfortunately, since none of
those fields are marked as optional, a user cannot customize the view
and decide to toggle those fields.

This commit adds the optional attribute on several fields to allow some
customization.

* concerned fields
- name
- tag_ids
- partner_id
- owner_id
- type
- create_date

Task-id 2846754

closes odoo/enterprise#27093

Signed-off-by: Alexandre Kühn (aku) <aku@odoo.com>