[REF] documents_spreadsheet: hide tag selector widget compatibility layer
This commit wraps in a component the compatibility layer of the tag selector widget.
The goal is to decouple this layer from the business code which should not kown anything
about such details. It will also reduce the impact when the widget and the required
compatibility layer will be removed.

closes odoo/enterprise#27130

Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>