[IMP] payment: rename of acquirer on provider
Chaning the name of model payment.acquirer to payment.provider
and everything that it touches. It is technically incorrect to
use the term "acquirer" for systems that only provide a service
of payment.
After this commit the model payment.acquirer and everything
related to it will be renamed to payment.provider.

Task - 2842088

closes odoo/enterprise#27131

Related: odoo/upgrade#3542
Related: odoo/documentation#1981
Related: odoo/odoo#90899
Signed-off-by: Victor Feyens (vfe) <vfe@odoo.com>