[IMP] project_enterprise,industry_fsm: link milestone to task
Before this commit, the milestones are only linked to a project without
any link to the tasks of the same project. However, the tasks could be
considered as steps to reach a project milestone.

This commit manages the `allow_milestones` field in project to fsm
project, by default the feature will not be enabled for the fsm
projects. It also adds the `milestone_id` field defined in
`project.task` model, in some views (gantt, for instance).

task-2829542