[IMP] helpdesk : measure response time
The following measures have been added to ticket analysis.

Hours to first response:
nb working hours between the date on which the ticket was received
and the date on which the first message was sent by the support
(exclude emails sent automatically through the stages).

Hours to respond:
avg nb of working hours between a message from the customer
and a message from the support
(only applicable for messages sent when the ticket was in a non folded stage).

task-2847937

closes odoo/enterprise#27158

Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>