[REF] *: Analytic Apocalypse
The rationale behind the analytic tags and accounts changes can be found in the community commit.

There are also changes to the reports:
The analytic report has been removed.

The analytic filter on financial reports (balance sheet and profit loss report) has been removed, to be replaced by an analytic groupby filter, that will create additional columns on the financial report, based on the analytic lines linked to the financial account of the line and the analytic account (or plan) of the column.

To do so, we use a temporary table that will be computed and filled once, with the values of the analytic lines, to shadow the table of account move lines. This lets us use all the engine defined for account move lines.

On all reports, not financial ones, that had an analytic filter will keep the old one.

closes odoo/enterprise#30743

Related: odoo/odoo#98914
Related: odoo/upgrade#3885
Signed-off-by: William André (wan) <wan@odoo.com>