[IMP] account_followup: make inheritable domain for unreconciled AMLs
Recently, partner field `unreconciled_aml_ids` was changed to a
computed one on 76dc8ed7 and its domain was moved from the field itself
to inside the compute method. Unfortunately, that made the domain to be
harcoded into that method.

This commit moves the domain to a separate method, so it may be
inherited or reused.

closes odoo/enterprise#27397

X-original-commit: 0e39dfd024dc22ad838e90d0f6dd76d4dfe7cb59
Signed-off-by: William André (wan) <wan@odoo.com>