[FIX] sale_subscription: increment date in an override of _create_invoices
Before this commit, the next_invoice_date increment was performed after the
call to _create_invoices on sale_order. When the option 'Automatic invoice'
was selected, when the customer accepted the quotation, two invoices were
created: one for the automatic flow and another one from the cron because
the next_invoice_date was not properly updated.

Part-of: odoo/enterprise#27311