[FIX] hr_appraisal: remove historically implicit manager rights
Before, if an employee A has ever been an appraisal manager for an
employee B, they will see all employee B's past and future appraisals.

But in real life, employee A (who is not employee B's manager) is probably
appraisal manager, cause they collaborated on a task or so.
Now, it's not because you collaborated for a project or a task for a given
period of time, that this collaboration will go on.

So, Employee B should only see the appraisal that he is manager on.

Same logic applies to appraisal goals. B should not be able to see A's
goals.

task - 2849301

closes odoo/enterprise#27214

Related: odoo/upgrade#3515
Signed-off-by: Kevin Baptiste <kba@odoo.com>