[FIX] stock_barcode: apply the destination package
On a picking, if a line already has a source package, it won't be
possible to define its destination package

To reproduce the issue:
1. In Settings, enable "Packages"
2. Create a storable product P
3. Update its qty:
    - 2 x P in package PACK01
4. Inventory > Products > Packages, create a second package PACK02
5. Create a planned delivery order DO with 1 x P
6. Mark DO as todo
    - One P is reserved (its source package is PACK01 but its
destination package is not defined)
7. Barcode > Operations > Delivery Orders, open DO
8. Click on +1 for P
9. Scan PACK02

Error: The destination package is not completed and a error notification
is displayed "You are expected to scan one or more [...]"

Processing the barcode leads to `_processPackage` because the barcode is
considered as a package, which is correct. However, it does not define
the destination package with PACK02 because the selected line already
has a source package. This should not matter.

OPW-2822494

closes odoo/enterprise#27562

X-original-commit: 5d3317b6e01ca694f283ddc1faaed50e40f18e26
Signed-off-by: Steve Van Essche <svs@odoo.com>
Signed-off-by: Adrien Widart <awt@odoo.com>