[IMP] payment_sepa_direct_debit: immediately confirm transactions
Before this commit, when the customer chose SEPA Direct Debit to make a
payment, the transaction was set in the `pending` state after creating
the mandate. An `account.payment` was immediately created to allow
accountants to order the bank debit. After reconciliation of the
payment, which can take weeks, the transaction was confirmed (set to
`done`) and post-processed. Only then were the related documents
(quotation, subscription, invoice) confirmed.

This way of doing posed several problems:
- Most apps consider the `pending` state as a near-failure. When the
  customer was being redirected to the portal page of the document they
  intended to pay, no mention was made of the pending SEPA payment.
  Worse, the customer was invited to pay again.
- As SEPA payments take a lot of time to process, the confirmation of
  the document to pay was significantly delayed. In the Subscriptions
  app, in particular, this poses problems because the subscription was
  not started or resumed even though the customer had provided all the
  required payment details.
- Creating the `account.payment` record *before* setting the transaction
  to the state `done` implied extra technical complexity because every
  other payment flow waits for the transaction to be confirmed before
  creating the related payment, which is the logical way to post-process
  transactions. Indeed, we should typically wait for the funds to be
  available for reconciliation to create the payment, which was not the
  case for SEPA payments.

These problems can be addressed thanks to the following observations:
- By law, the customer must be able to cancel SEPA payments
  effortlessly. Therefore, handling the case where the customer cancels
  the payment after the paid document has been confirmed is required.
- SEPA mandates work conceptually like payment tokens: after they are
  created, one can use them to make payments without the intervention of
  the customer. A SEPA payment can therefore be considered as "almost
  confirmed" as soon as the mandate is created.

After this commit, SEPA payments are confirmed (the state is set to
`done`) as soon as the mandate is created. This avoids having to deal
with all the side effect of the `pending` state altogether and resolves
the technical difficulties related to the timing of the creation of the
`account.payment`. Should a SEPA payment fail, handling it would be the
same process as handling a payment cancelation.

task-2842241

closes odoo/enterprise#27251

Signed-off-by: Antoine Vandevenne (anv) <anv@odoo.com>