[FIX] account_intrastat: supplementary units
The intrastat declaration now requires supplementary units. The
supplementary units are used when they are more relevant than the
default weight in kilograms for the commodity code of the line.

Because this commit is in stable, it utilises a python dictionary for
the relation between supplementary units and commodity codes.
This dictionary contains a mapping for only those commodity codes which
have supplementary units.

A function is added to the intrastat report that maps the data on the
units of measure in the results from the query, to the appropriate
supplementary unit for the report. The function is also applied in the
cbs export in l10n_nl_intrastat cbs export in order to fill the correct
section of the fixed-position-file 'SRL' file that gets uploaded to the
dutch authority (cbs).

Also included are tests to test the expected behavior of these units
from invoice to report.

closes odoo/enterprise#28149

Task-id: 2853030
X-original-commit: 0c82beaf7ccf55391a0a788489029f2b3fbd7fce
Signed-off-by: Josse Colpaert <jco@odoo.com>
Signed-off-by: Daniel Kosky (dako) <dako@odoo.com>