[FIX] account_avatax: create taxes with the right country_id
default_get() of account.tax uses self.env.company to determine the
default country_id.

If a user belongs to a company other than the company of the invoice
then the new tax is in the right company but the wrong country leading
to:

  File ".../odoo/models.py", line 1367, in _validate_fields
    check(self)
  File ".../odoo/addons/account/models/account_move.py", line 2099, in _validate_taxes_country
    raise ValidationError(_("This entry contains some tax from an
    unallowed country. Please check its fiscal position and your tax
    configuration."))

    odoo.exceptions.ValidationError: This entry contains some tax from an
    unallowed country. Please check its fiscal position and your tax
    configuration.

This could happen when the public user renewed via
/my/subscription/<id> and the public user was in a different company
than the subscription and resulting invoice.

task-2824272

closes odoo/enterprise#27413

X-original-commit: 388e96bb9b567a15b097de9ef9aa8c1416e9b90e
Signed-off-by: William André (wan) <wan@odoo.com>