[FIX] sale_subscription: fallback client order reference to name
Currently the subscriptions' client order references default to false, as this
is the behavior for all sale orders. However we still use the order reference
in emails, logs and even for searches, resulting in empty words in logs/emails
and searching using a False domain on payment. This commit fallbacks these to
the subscription name.

task-2824236

X-original-commit: 5fb66e97d1eef7fe29eaf8a2cac1778acb9c69b7
Part-of: odoo/enterprise#27750