[IMP] helpdesk, account: compute the KPI in batch
Purpose
=======
Now, the KPI are computed based on their `company_id` and not based
on the current company. If no company is set on the digest, we compute
it based on the current company.

The KPIs are computed based on their company. Most of the time, they
will all belong to the same company so we can improve the performance
by computing them in batch.

Task-2827996
See odoo/odoo/pull/91945

closes odoo/enterprise#27658

Related: odoo/odoo#91945
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>