[IMP] mail_enterprise, *: replace widget el by target
*: account_invoice_extract, documents.

This PR prepares the ground for the one introducing the new environment in the
discuss app. Indeed, the createWebClient helper will be used instead of createView/Widget
thus the returned value won't be a Widget anymore but a DOM element. This change impacts
they way the tests were written: JQuery functions have to be transformed to native JavaScript
ones. In order to reduce the noise in the main PR, those changes are done now.

task-2582313

closes odoo/enterprise#27313

Related: odoo/odoo#91290
Signed-off-by: Sébastien Theys (seb) <seb@odoo.com>