[IMP] (sale_)planning : generic improvements
Planning > Filters :
Add shifts in overtime.

Planning > Plan Orders > Succes notification :
Add a button 'View Shifts' to see the orders just planned.

Planning > Tour :
Rainbow man should be displayed when the user is back
on the planning view, after publishing it.
Yet when a resource is added during the tour, a modal pops
to add an employee and rainbow man pops behind it.
To solve that, and to enrich the tour, 2 more steps were added:
see next period and copy previous

task-2851381

closes odoo/enterprise#27312

Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>