[FIX] stock_barcode: avoid useless line creation
In some cases, the option "Create New Lots/Serial Numbers" is not
respected

To reproduce the issue:
1. Create a tracked-by-lot product P
2. Update the quantities:
    - 1 x P01 with lot LOT01
3. Open the operation type "Delivery Orders":
    - "Create New Lots/Serial Numbers" must be disabled
4. Barcode > Operations > Delivery Orders, Create
5. Scan LOT01
6. Scan ANYTHING

Error: a new line is created for the lot "ANYTHING". Because of step 3,
scanning the lot should not do anything.

(Side note: this commit also fixes a typo, `barcode` is a string, we
should use `barcodeData` instead)

OPW-2798144

X-original-commit: 6a3a97c75db182e345a9b661b0d89d5c941dd815
Part-of: odoo/enterprise#27769