[FIX] stock_barcode: scan lots/SN in a new delivery
Suppose a new delivery (without any move). Scanning some tracked
products does not work.

To reproduce the issue:
1. Create two tracked-by-lot products P01, P02 and a tracked-by-sn one
P03:
    - P01:
        - Barcode: 0010000000009
    - P02:
        - Barcode: 0020000000008
2. Update the quantities:
    - 1 x P01 with lot L01 (name: "SUPERLOT")
    - 1 x P02 with lot L02 (name: "SUPERLOT")
    - 1 X P03 with serial number SN (name: "SUPERSN")
3. Open the operation type "Delivery Orders":
    - "Create New Lots/Serial Numbers" must be disabled
4. Barcode > Operations > Delivery Orders, Create
5. Scan 0010000000009 (i.e., P01)
6. Scan SUPERLOT
7. Scan 0020000000008 (i.e., P02)
8. Scan SUPERLOT
    - Error01: The quantity of P01 has been updated, it is now equal to
2 and the quantity of P02 is still zero. This is incorrect. We should
have 1 x P01 and 1 x P02
9. Scan SUPERSN
    - Error02: A notification is displayed ("You are expected to scan
one or more products."). However, SUPERSN exists, a new line for P03
should be added

Error01:

Because the selected line (line of P02, thanks to step 7) is tracked by
lot, we add a filter in case the barcode is a lot:
https://github.com/odoo/enterprise/blob/a680071e2da1c1198065e7ac706f77a773787f4f/stock_barcode/static/src/models/barcode_model.js#L1091-L1096
So, we expect that the product of the lot will be P02. Later on,
`getRecordByBarcode` is called to get the record related to the barcode.
In the for-loop, we find some `ids` for the model
`stock.production.lot`. This is actually the lot added at step 6 (L01).
We directly add this lot in the `result` and then check if the record
respects the filters (so we check if the product of L01 is P02). Because
the condition is not respected, and because `ids` does not contain any
other identifier, the for-loop is stopped, but `result` still contains
L01 -> this is the issue, it explains why the wrong quantity is updated.

Once the above issue is fixed, we need to fix two other issues
(otherwise the above use case still does not work):
- Because `result` is now empty, we try to find the missing record
thanks to `_getMissingRecord`. However, this method has already been
called at step 5 for the same barcode ("SUPERLOT"). As a result, the
barcode has been saved in `missCache`. Back to step 7, it will lead to:
https://github.com/odoo/enterprise/blob/5b29cf6d96bc26139c767188fb68f3fa56a1a537/stock_barcode/static/src/lazy_barcode_cache.js#L156-L158
i.e., we will consider that we have already looked for that barcode and
the we don't need to look for it again. We should not complete
`missCache` when there are some filters
- Once `_getMissingRecord` is processed, we call again
`getRecordByBarcode` to recompute `result`:
https://github.com/odoo/enterprise/blob/5b29cf6d96bc26139c767188fb68f3fa56a1a537/stock_barcode/static/src/lazy_barcode_cache.js#L138-L139
However, because we don't provide the filters anymore, it will find and
return the lot L01. We should provide the filters.

Error02:

The selected line is still the line for product P02. Therefore, as
above, we add a filter: "product must be P02". This time, in
`getRecordByBarcode`, it won't find any `ids`. It will then send a RPC.
However, because of the domain, it won't find anything. That's the
reason why a notification is displayed.

When calling `_parseBarcode` with some filters, if there isn't any
result, we should try again without the filter on the product.

OPW-2798144

closes odoo/enterprise#27769

X-original-commit: 7dc7d72f45b483cb0336f4ad7ed72d0d28fcb5e5
Signed-off-by: Steve Van Essche <svs@odoo.com>
Signed-off-by: Adrien Widart <awt@odoo.com>