[IMP] stock_barcode: detailed backorder wizard
If `display_detailed_backorder` is in the context, the backorder wizard
will display a different message, counting how many moves are not
processed (totally or partially).
It's usefull for the Barcode App because if the user tries to validate a
uncomplete picking, they will directly know how many moves weren't fully
processed (it's harder to see in the Barcode App because it can have
multiple pages, one by location, and the uncomplete moves can be
splitted between different pages).

task-2666478

Part-of: odoo/enterprise#27547