[FIX] l10n_cl_edi_*: replace non iso-8859 chars
The XML for the electronic invoice in CL needs to be
encoded in iso-8859, while the qweb renders everything in utf-8
and Odoo stored product names, partner names, ... in utf-8.

For example, a character like € is not in iso-8859, but in utf-8.

If you put that in the product description, it would now give
a traceback because it can not encode it as utf-8.  However,
if we put the replace, it wil at least not traceback
and we will at least see a (wrong, but doubtful) character:

Coin with €-sign -> Coin with ?-sign.

opw-2844036

closes odoo/enterprise#28914

X-original-commit: 8caabde085ff2403fb941d49e0cbede993e6313c
Signed-off-by: William André (wan) <wan@odoo.com>