[IMP] wesbite_appointment: handle custom background color
PURPOSE:

When using a custom background color, it is now used everywhere in
website appointment.

To achieve this behavior and to maintain good readability, some
components have now a style that depends on the theme's colors, rather
than having hard-coded colors.

For example, the appointment selection view had a grey background, no
matter the theme's background color, so that the white cards stood out
of this background. The background color is now the one of the theme,
and the shadow of the cards is now a little bit stronger so that the
cards still stand out on a white background. Cards are still white, to
be consistent with the cards of the columns snippet.

Task-2835882

closes odoo/enterprise#26789

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>