[IMP] appointment: trigger on change while typing the short code
Currently while sharing an appointment, there is some misbehavior while
manually changing the code:

1/ clicking on "Save & Copy" will copy the old link (before manual change,
   unless another click is done first).
2/ if you type an already used short code, and immediately click on
   "Save & Copy", the rpc call is triggered and raises the validation error
   (before the div alert is visible, and the button gets disabled).

Both of them happens because the onchange on the short code won't trigger
unless we click somewhere else before hitting the "Save & Copy" button.

With this commit, we enable triggering the on change while typing, with a
short debounce delay of 0.5 seconds. Apart from that, we now do not hide
the whole URL if short code is emptied by user.

task-2821978

closes odoo/enterprise#28822

X-original-commit: 1011733bc33f56818ec47124f608c3438628dc78
Related: odoo/odoo#94486
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>