[FIX] sign: wrong displayed checkbox and reset placeholder
Steps to reproduce:
Install sign and activate developper mode
Set an empty placeholder on the "Text" field type
Go to a sign template and add a Text field
Make this field not mandatory and validate

Expected behavior:
The placeholder remain false

Current behavior:
The placeholder is set to Text

Explanation:
In the code when the value of the name was empty
then it was reset to the field name. In this solution
the value saved in the name of sign_item is false
if there is no placeholder and the placeholder otherwise.
Previously the value false of this field could mean the
default placeholder for this field type or an empty string.

opw-2767417

closes odoo/enterprise#28234

X-original-commit: cbda7bdc2ebaccb708fffc3c7624c6f023a09f56
Signed-off-by: Arnaud Joset <arj@odoo.com>
Signed-off-by: Fockedey Martin (mafo) <mafo@odoo.com>