[FIX] mrp_plm: compare the difference between two BOMs
Steps to reproduce the product:
- Enable “Work Orders” option in MRP settings
- Create a storable product “P1” with BOM:
    - BOM type: Manufacture this product
    - operations: “OP1”
    - Components: product “C1” consumed in the operation “OP1

- Create an ECO for “P1”:
    - Type: New Product Introduction
    - Apply on; Bill Of materials
    - Click on “Start Revision”

Problem:
A BOM change will be created for the component “C1” where it will be added and removed.

When creating an ECO, a copy of the original BOM will be created,
and then we compare the BOM lines of the two BOMs if they are identical (product, qty and operation)
but as the operation has been duplicated therefore created with a different ID,
the two lines will not be identical and a distorted change in the BOM will be detected

Solution:
- It is better to compare the fields of the operations instead of their ID to detect if a change has been made in the BOM

- When modifying the revision BOM so that a component is consumed in another operation than the original one,
a BOM line changes type “Remove” and “Add” will be added,
because we cannot detect that an update has been made, imagine we have the following BOM:

1:/ old_bom to produce “product A”:

Components:

“Product B” - 1 unit - operation 1 (id = 1)
“Product B” - 1 unit - operation 2 (id= 2)
-------------------------------------------------- -----------------

2:/ new_bom == copy of (old_bom)

Components:

Product B - 1 unit - Operation 1 (id = 3)
Product B - 1 unit - operation 2 (id= 4)

-------------------------------------------
Operation 1 (id=1) == Operation 1 (id=3)
Operation 2 (id = 2) == Operation (id = 4)

So as the product is the same in the two BOM lines, if we change the two operations at the same time,
we cannot check which line exactly was changed.

opw-2823253
opw-2809586
opw-2833083

closes odoo/enterprise#27943

X-original-commit: 9f697161bac4cdf0cb53e2c996a0fe8c5eab139e
Related: odoo/odoo#92670
Signed-off-by: Djamel Touati (otd) <otd@odoo.com>
Signed-off-by: William Henrotin (whe) <whe@odoo.com>