[FIX] sale_project_forecast: ensure project_id is updated accordingly
When generating a task in a project from a sale line order, the task and project are
created and the `task_id` and `project_id` are set on the sol. If one changes the `project_id`
of the task, the constraint check made in `_check_task_in_project` in `planning.slot` will fail.

Prior to this fix:
 - When the task project_id is changed, `planning.slot` `_compute_project_id` is triggerd thanks to its
   depends on `sale_line_id.task_id.project_id` but the value taken will be `sale_line_id.project_id`
   which does not reflect the changes of the project_id on the task.

After this fix:
 - `_compute_project_id` will preferably take `slot.sale_line_id.task_id.project_id`
   over `slot.sale_line_id.project_id` which will prevent `_check_task_in_project` to
   fail.

task-2854845

closes odoo/enterprise#27464

X-original-commit: 8a6c291595085d4cbbbc5c0bb12daffd38abca51
Signed-off-by: Xavier <xbo@odoo.com>
Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>