[IMP] mail_enterprise: allow ui size patching during tests
Until now, when we add to patch the ui size, we would have patched one of the following:
       - config.device.{isMobile,size_class}
       - env.isSmall / env.services.ui.{isSmall/size}
       - browser.innerWidth / window.innerWidth

The issue is that patching one does not affect the others resulting in an incoherent state.
In order to ease ui size patching, the `patchUiSize` method has been introduced. This method
take either a size or a width as parameter (eg. SIZES.SM, 1920) and affect all the properties above.

task-2582313

closes odoo/enterprise#27428

Related: odoo/odoo#91540
Signed-off-by: Sébastien Theys (seb) <seb@odoo.com>