[FIX] account_asset : fix asset modification with prorata
To reproduce the issue:
Create an asset with the prorata option checked and a value of 1000 and 3 depreciations, then modify the asset with the same values (3 depreciations). An additional entry (4/3) is created while it shouldn't.

To correct the issue:
We need to use the context information that we should ignore the prorata in the board computation.

closes odoo/enterprise#27573

X-original-commit: 12f43d29f659eac374a758044ab6e3ea458a4601
Signed-off-by: William André (wan) <wan@odoo.com>