[FIX] documents: fix the canvas rendering after rasterization
Since the update to OWL2, the part of the code that was adding the
canvas to the DOM was removed. This part was a trick to individually
re-render pdfPage components after the pdf.js library finished
the rasterization of that page.

This commit moves the pageCanvases object to the state to use the normal
OWL rendering mechanism to render the pages, which was not possible
before OWL2.

closes odoo/enterprise#27445

X-original-commit: c3d696f99ad74b9de3035e9e24f9734535871c0e
Signed-off-by: Alexandre Kühn (aku) <aku@odoo.com>