[FIX] timesheet_grid: prevent timesheet self validation - approvers
Steps to reproduce:
- Create a user/employee
- Give Timesheet rights: approver
- Create a timesheet for the user/employee

Issue:
When connected with this user, he can self-approve his own timesheet

Cause:
It is currently specifically designed that way:
"""Approver: in this access right, the user can't validate all timesheets,
he can validate the timesheets where he is the manager or timesheet responsible of the
employee who is assigned to this timesheets or the user is the owner of the project.
Furthermore, the user can validate his own timesheets."""

Nonetheless, according to the team and the PO (aju) it has to be changed.
=> An approver can not self approve his own timesheets.

Solution:
Restrict the domain to take the lines where the employee_id is not the same as the one from the user validating.

opw-2836385

closes odoo/enterprise#27829

X-original-commit: 572b874493f59108fd1e361d0036d4e5bc06ad2e
Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>